Create 07_usercreds (#39)
* Create 07_usercreds

user credentials for all non-dummy sample data. (the dummy I made in 06 is excluded as I want the dummies to remain unchanged for demo purposes). Tested fine on my end so if it works for others as well, should be fine for merge.

* Rename 07_usercreds to schema/07_usercreds

changed the file path to the schema folder for improved organization